+++ flutter_app/lib/screens/log_displayer_screen.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class LogDisplayerScreen extends StatelessWidget{
+  final String fileContext;
+  const LogDisplayerScreen ({Key? key, required this.fileContext}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Log Displayer'),
+      ),
+      body: ListView(
+        children: [
+          Text(fileContext),
+        ],
+      ),
+    );
+  }
+}
